Combine non-AMR and AMR multilayer test
Also contains a new set of regression data.  The previous regression data
had a buggy version of the friction coefficient setting.  This will hopefully
reconcile all the changes to provide a better testing basis for the new AMR
code.
